Refactor two excel exporter, #9892
